Unset backport branch's remote from upstream
This way the user can easily commit and push further changes to the backport,
if needed, without having to remember to append their fork's remote name to
`git push`.

The branch sets the upstream as the remote during `git checkout -b`. This isn't
a useful default as we push the resulting backport to the user's fork anyway
and it is never intended to either merge further upstream changes nor to push
to upstream directly.